+++ src/libfpdec/digit_array.h
@@ -41,7 +41,7 @@ digits_copy(const fpdec_digit_array_t *src, fpdec_n_digits_t n_shift,
 error_t
 digits_from_dec_coeff_exp(fpdec_digit_array_t **digit_array, fpdec_exp_t *exp,
                           size_t n_dec_digits, const dec_digit_t *coeff,
-                          int dec_exp);
+                          int64_t dec_exp);
 
 error_t
 digits_from_digits(fpdec_digit_array_t **digit_array,
